# api/openapi.yaml
openapi: 3.0.0
info:
  title: WatchWorld API
  version: 1.0.0
  description: API that lives alongside the SQL Server in Docker.
servers:
  - url: http://localhost:8080/api
    description: Local (Docker)
paths:
...
  /Watches:
    get:
      description: Returns all watches
      responses:
        '200':
          description: OK
    post:
      description: Creates a new watch
      responses:
        '201':
          description: Created
        '400':
          description: Invalid request
  /Watches/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      description: Returns one watch
      responses:
        '200':
          description: OK
        '404':
          description: Not found
    put:
      description: Updates a watch
      responses:
        '200':
          description: Updated
        '400':
          description: Invalid request
        '404':
          description: Not found
    delete:
      description: Deletes a watch
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found


  # ---------------- IndividualWatch ----------------
  /IndividualWatch:
    get:
      description: Returns all individual (physical) watches
      responses:
        '200':
          description: OK
    post:
      description: Creates an individual watch
      responses:
        '201':
          description: Created
        '400':
          description: Invalid request
  /IndividualWatch/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      description: Returns one individual watch
      responses:
        '200':
          description: OK
        '404':
          description: Not found
    put:
      description: Updates an individual watch
      responses:
        '200':
          description: Updated
        '400':
          description: Invalid request
        '404':
          description: Not found
    delete:
      description: Deletes an individual watch
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found

  # ---------------- Listing ----------------
  /Listing:
    get:
      description: Returns all listings
      responses:
        '200':
          description: OK
    post:
      description: Creates a listing
      responses:
        '201':
          description: Created
        '400':
          description: Invalid request
  /Listing/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      description: Returns one listing
      responses:
        '200':
          description: OK
        '404':
          description: Not found
    delete:
      description: Deletes a listing
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found

  # ---------------- Borrow ----------------
...